FIX: SQL 쿼리 수정

+++ migrations/005_create-views.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS views (
-    thread_id INTEGER NOT NULL,
+    thread_id BIGINT NOT NULL,
     view_count INTEGER DEFAULT 0,
     
     PRIMARY KEY(thread_id),
